w3c-reports/12092014/efl/runned-tests/dom/ranges/Range-cloneContents-expected.txt

CONSOLE MESSAGE: line 257: TypeError: undefined is not a function (evaluating 'iframe.contentWindow.setupRangeTests()')
FAIL: Timed out waiting for notifyDone to be called
